+++ Makefile
@@ -1,13 +1,13 @@
 VERSION=$(shell git describe --tags)
 
 build:
-	go build -ldflags "-X main.version=$(VERSION)"
+	go build -x -ldflags "-X github.com/pocke/lemonade/lemon.Version=$(VERSION)"
 
 install:
-	go install -ldflags "-X main.version=$(VERSION)"
+	go install -ldflags "-X github.com/pocke/lemonade/lemon.Version=$(VERSION)"
 
 release:
-	gox --arch 'amd64 386' --os 'windows linux darwin' --output "dist/{{.Dir}}_{{.OS}}_{{.Arch}}/{{.Dir}}" -ldflags "-X main.version=$(VERSION)"
+	gox --arch 'amd64 386' --os 'windows linux darwin' --output "dist/{{.Dir}}_{{.OS}}_{{.Arch}}/{{.Dir}}" -ldflags "-X github.com/pocke/lemonade/lemon.Version=$(VERSION)"
 	zip      pkg/lemonade_windows_386.zip     dist/lemonade_windows_386/lemonade.exe   -j
 	zip      pkg/lemonade_windows_amd64.zip   dist/lemonade_windows_amd64/lemonade.exe -j
 	tar zcvf pkg/lemonade_linux_386.tar.gz    -C dist/lemonade_linux_386/    lemonade
